deeplabv3/requirements.txt
absl-py==1.4.0
accelerate==0.27.2
acvl-utils==0.2
addict==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
batchgenerators==0.25
beartype==0.17.2
blessed==1.20.0
cachetools==5.3.0
# certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==2.1.1
click==8.1.7
cloudpickle==2.2.1
cmake==3.25.0
connected-components-3d==3.10.5
contextlib2==21.6.0
contourpy==1.0.7
cycler==0.11.0
dicom2nifti==2.4.8
docopt==0.6.2
dynamic-network-architectures==0.2
efficientnet-pytorch==0.7.1
einops==0.6.0
et-xmlfile==1.1.0
filelock==3.9.0
flatbuffers==23.3.3
fonttools==4.39.2
frozenlist==1.4.1
fsspec==2024.2.0
future==0.18.3
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpustat==1.1.1
grad-cam==1.4.8
graphviz==0.20.1
grpcio==1.51.3
h5py==3.8.0
# Editable Git install with no remote (HD-BET==1.0)
-e /mai_nas/BYS/brain_metastasis/HD-BET
huggingface-hub==0.13.2
idna==3.4
imagecodecs==2023.3.16
imageio==2.26.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
jax==0.4.6
Jinja2==3.1.2
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.1
libclang==16.0.0
lightning-utilities==0.10.1
linecache2==1.0.0
lit==15.0.7
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
mc-bin-client==1.0.1
MedPy==0.4.0
ml-collections==0.1.1
mmcv==1.7.1
monai-weekly==1.2.dev2320
mpmath==1.2.1
multidict==6.0.5
munch==4.0.0
natsort==8.3.1
networkx==3.0
nibabel==5.1.0
ninja==1.11.1
Editable Git install with no remote (nnunet==1.7.0)
-e /mai_nas/BYS/nnUNet
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==12.550.52
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.1
Pillow==9.5.0
pretrainedmodels==0.7.4
protobuf==4.22.1
psutil==5.9.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydicom==2.3.1
pykwalify==1.8.0
pyparsing==3.0.9
pyradiomics==3.0.1
python-dateutil==2.8.2
python-gdcm==3.0.22
pytorch-lightning==2.2.0.post0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
safetensors==0.3.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.12.2
segmentation-models-pytorch==0.3.3
SimpleITK==2.2.1
six==1.16.0
submitit==1.5.0
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
tifffile==2023.3.15
timm==0.5.4
# torch==1.13.1
# torchaudio==0.13.1+cu117
torchmetrics==1.3.1
# torchvision==0.14.1+cu117
tqdm==4.65.0
traceback2==1.4.0
triton==2.0.0
ttach==0.0.3
typing_extensions==4.5.0
tzdata==2023.3
unittest2==1.1.0
urllib3==1.26.15
wcwidth==0.2.13
Werkzeug==2.2.3
wrapt==1.14.1
yacs==0.1.8
yapf==0.32.0
yarl==1.9.4
zipp==3.15.0
